Applied critical/non-critical post review fixes
