+++ ImageFeed/Auth/URLSession+data.swift
@@ -7,19 +7,6 @@
 
 import Foundation
 
-enum NetworkError: Error {
-    case authorizationRequiredError
-    case webPageNotFoundError
-    case internalServerError
-    case urlSessionError
-}
-
-enum HTTPStatusCode: Int {
-    case authorizationRequiredStatus = 401
-    case webPageNotFoundStatus = 404
-    case internalServerErrorStatus = 500
-}
-
 extension URLSession {
     /// Loads the web data on the main queue
     func data( for request: URLRequest, completion: @escaping (Result<Data, Error>) -> Void ) {
@@ -34,14 +21,14 @@ extension URLSession {
                 let data = try await self.loadNetworkData(request: request)
                 fulfillCompletionOnTheMainThread(.success(data))
             } catch {
-                print("Failed to loadNetworkData request")
+                print("Failed to loadNetworkData request error code \(error)")
                 fulfillCompletionOnTheMainThread(.failure(error))
             }
         }
     }
     
     /// Checks for response code from the server
-    func loadNetworkData(request: URLRequest) async throws -> Data {
+    private func loadNetworkData(request: URLRequest) async throws -> Data {
         let (data, response) = try await URLSession.shared.data(for: request)
         if let response = response as? HTTPURLResponse,
            response.statusCode < 200 || response.statusCode >= 300
@@ -53,7 +40,7 @@ extension URLSession {
     }
     
     /// Handles server's error responses
-    func showNetworkError(for statusCode: Int) -> NetworkError {
+    private func showNetworkError(for statusCode: Int) -> NetworkError {
         if let httpsCode = HTTPStatusCode(rawValue: statusCode) {
             switch httpsCode {
             case .authorizationRequiredStatus:
