+++ ImageFeed/Base.lproj/LaunchScreen.storyboard
@@ -17,11 +17,11 @@
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
                             <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="splash_screen_logo" translatesAutoresizingMaskIntoConstraints="NO" id="C81-jb-ktB">
-                                <rect key="frame" x="160.33333333333331" y="401" width="73" height="75"/>
+                                <rect key="frame" x="160" y="401" width="73" height="75"/>
                             </imageView>
                         </subviews>
                         <viewLayoutGuide key="safeArea" id="6Tk-OE-BBY"/>
-                        <color key="backgroundColor" red="0.10196078431372549" green="0.10588235294117647" blue="0.13333333333333333" alpha="0.0" colorSpace="custom" customColorSpace="displayP3"/>
+                        <color key="backgroundColor" name="YP_black_color"/>
                         <constraints>
                             <constraint firstItem="C81-jb-ktB" firstAttribute="centerY" secondItem="6Tk-OE-BBY" secondAttribute="centerY" id="6dc-yN-MvD"/>
                             <constraint firstItem="C81-jb-ktB" firstAttribute="centerX" secondItem="6Tk-OE-BBY" secondAttribute="centerX" id="rVK-z3-fpI"/>
@@ -35,5 +35,8 @@
     </scenes>
     <resources>
         <image name="splash_screen_logo" width="72.666664123535156" height="75.333335876464844"/>
+        <namedColor name="YP_black_color">
+            <color red="0.10196078431372549" green="0.10588235294117647" blue="0.13333333333333333" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+        </namedColor>
     </resources>
 </document>
