+++ ImageFeed/Models/Constants.swift
